+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.1.2-dev
 
 - Send `fetch` requests instead of `XHR` requests.
+- Add an optional `debugKey` parameter to `SseClient` to include in logging.
 
 ## 4.1.1